refactor: use Decimal

+++ backend/libraries/ballerina-core-python/ballerina_core/parsing/parsing_types.py
@@ -1,9 +1,10 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Mapping, Sequence
+from decimal import Decimal
 from typing import TypeVar
 
-Json = Mapping[str, "Json"] | Sequence["Json"] | str | int | float | bool | None
+Json = Mapping[str, "Json"] | Sequence["Json"] | str | int | Decimal | bool | None
 
 _ToJsonType = TypeVar("_ToJsonType")
 ToJson = Callable[[_ToJsonType], Json]
